vibe coded routing and decap content type
